Add a helper class for common wrapper methods
