+++ src/Transport/Dispatcher.cs
@@ -3,7 +3,6 @@ namespace NServiceBus.Transport.AzureStorageQueues
     using System;
     using System.Collections.Concurrent;
     using System.Collections.Generic;
-    using System.IO;
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
@@ -11,6 +10,7 @@ namespace NServiceBus.Transport.AzureStorageQueues
     using global::Azure.Storage.Queues;
     using Logging;
     using NServiceBus.AzureStorageQueues;
+    using NServiceBus.Transport.AzureStorageQueues.Utils;
     using Transport;
     using Unicast.Queuing;
 
@@ -124,16 +124,7 @@ namespace NServiceBus.Transport.AzureStorageQueues
 
         Task Send(MessageWrapper wrapper, QueueClient sendQueue, TimeSpan timeToBeReceived, CancellationToken cancellationToken)
         {
-            string base64String;
-
-            using (var stream = new MemoryStream())
-            {
-                serializer.Serialize(wrapper, stream);
-
-                var bytes = stream.ToArray();
-                base64String = Convert.ToBase64String(bytes);
-            }
-
+            string base64String = MessageWrapperHelper.ConvertToBase64String(wrapper, serializer);
             return sendQueue.SendMessageAsync(base64String, timeToLive: timeToBeReceived, cancellationToken: cancellationToken);
         }
